pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "bcrypt==4.3.0",
    "fastapi[standard]>=0.128.0",
    "jwt>=1.4.0",
    "passlib>=1.7.4",
    "psycopg2>=2.9.11",
    "python-jose>=3.5.0",
]
